requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
pandas>=2.1.0
nvdlib>=0.7.4
loguru>=0.7.0
python-dotenv>=1.0.0
pathlib==1.0.1
flask==3.0.2 